kt编译新增特殊方法（_clearFindViewByIdCache）；支持多文件传输；支持传输规则文件

+++ src/main/kotlin/com/example/viewdebugtrans/AdbSendAction.kt
@@ -6,9 +6,14 @@ import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.PlatformDataKeys
 import com.intellij.openapi.editor.impl.EditorImpl
 import com.intellij.openapi.fileEditor.FileDocumentManager
+import com.intellij.openapi.module.ModuleManager
+import com.intellij.openapi.module.ModuleUtil
 import com.intellij.openapi.project.ProjectManager
+import com.intellij.openapi.roots.CompilerModuleExtension
+import com.intellij.openapi.roots.ProjectRootManager
 import com.intellij.openapi.ui.Messages
 import java.io.File
+import java.util.LinkedList
 
 
 /**
@@ -42,17 +47,64 @@ processRequestF.invoke(tasktI, location)
  */
 
 class AdbSendAction(private val device: String) : AnAction(device) {
+    private val sendAction = LinkedList<FileItem>()
     override fun actionPerformed(e: AnActionEvent) {
         try {
+            sendAction.clear()
             ShowLogAction.builder.clear()
             val editor = e.getData(PlatformDataKeys.EDITOR)
+            val project = e.project ?: return
             if (editor is EditorImpl) {
                 var path = FileDocumentManager.getInstance().getFile(editor.document)?.path ?: return
-                val makeRClass = MakeRClass()
-                makeRClass.make(e, path)
-                // 经过编译的产物路径
-                path = CompileFileAndSend().compile(path, e)
-                makeRClass.delete()
+                if (path.endsWith(".java") || path.endsWith(".kt")) {
+                    // 代码文件，需要编译和R文件
+                    val makeRClass = MakeRClass()
+                    makeRClass.make(e, path)
+                    // 经过编译的产物路径
+                    path = CompileFileAndSend().compile(path, e)
+                    makeRClass.delete()
+                } else if (path.endsWith(".xml")) {
+                    val rulePathsSet = HashSet<String>()
+                    val logSet = LinkedHashSet<String>()
+                    ModuleManager.getInstance(project).modules.forEachIndexed { index, it ->
+                        //val path = CompilerModuleExtension.getInstance(it)?.compilerOutputPath?.path
+                        val basePath = CompilerModuleExtension.getInstance(it)?.compilerOutputPath?.path?.replace('\\','/')
+                        if (basePath != null) {
+                            val index = basePath.indexOf("/build/")
+                            if (index > 0) {
+                                // xml文件，需要xml规则文件
+                                val rulesPath =  basePath.substring(0, index) + "/build/intermediates/incremental"
+                                val ruleFileDir = File(rulesPath)
+                                if (ruleFileDir.exists()) {
+                                    // 选择对应merge文件夹，过滤AndroidTestResources类型文件夹
+                                    val folder = ruleFileDir.listFiles()?.find { it.name.startsWith("merge") && it.name.endsWith("Resources") && !it.name.endsWith("AndroidTestResources") }
+                                    if (folder != null) {
+                                        // 需要设置不同的名称
+                                        val ruleFile = File(folder, "merger.xml")
+                                        if (ruleFile.exists()) {
+                                            // 过滤相同文件
+                                            if (!rulePathsSet.contains(ruleFile.absolutePath)) {
+                                                rulePathsSet.add(ruleFile.absolutePath)
+                                                // 推送规则文件
+                                                show(project, "找到规则文件-----：$ruleFile")
+                                                pushFile(ruleFile.absolutePath, Config.getTargetFileDestPath() + "merger-${index}.xml", device, "rules")
+                                            }
+                                        } else {
+                                            logSet.add("没有规则文件：$ruleFile")
+                                        }
+                                    } else {
+                                        logSet.add("没有规则文件：$rulesPath")
+                                    }
+                                }
+                            }
+
+                        }
+                    }
+                    logSet.forEach {
+                        show(project, it)
+                    }
+
+                }
 
                 val target = File(path)
                 if (target.exists()) {
@@ -68,6 +120,7 @@ class AdbSendAction(private val device: String) : AnAction(device) {
                         show(e.project!!, result)
                         Messages.showDialog(e.project, "推送成功", "提示", arrayOf("确定"), 0, null)
                     }
+                    pushApply()
 
                     //showDialog(editor.component)
                 } else {
@@ -89,16 +142,34 @@ class AdbSendAction(private val device: String) : AnAction(device) {
         return String(Runtime.getRuntime().exec(cmd).inputStream.readBytes())
     }
 
-    private fun pushFile(target: String, dest: String, device: String): String {
-        Config.saveConfig(dest)
+    private fun pushFile(target: String, dest: String, device: String, type: String = "file"): String {
+        // Config.saveConfig(dest, type)
+        addFileItem(dest, type)
         // 先推送文件
-        execute("adb -s $device push $target $dest")
+        return execute("adb -s $device push $target $dest")
         // 再推送config文件
-        return execute("adb -s $device push \"${Config.getConfigFile().absolutePath}\" ${Config.getConfigRemotePath()}")
+
+        // return execute("adb -s $device push \"${Config.getConfigFile().absolutePath}\" ${Config.getConfigRemotePath()}")
     }
 
 
     private fun checkRemoteFolder(device: String, folder: String) {
         execute("adb -s $device shell mkdir \"$folder\"")
     }
+
+    /**
+     * 推送配置文件
+     */
+    private fun pushApply() {
+        if (sendAction.isNotEmpty()) {
+            Config.saveConfig(sendAction)
+            execute("adb -s $device push \"${Config.getConfigFile().absolutePath}\" ${Config.getConfigRemotePath()}")
+        }
+    }
+
+
+    fun addFileItem(path: String, type: String) {
+        sendAction.add(FileItem(path, type))
+    }
+    class FileItem(val path: String, val type: String)
 }
